remove duplicated functions
